+++ x/apps/types/querier_test.go
@@ -8,69 +8,6 @@ import (
 	"testing"
 )
 
-func TestNewQueryStakedApplicationsParams(t *testing.T) {
-	type args struct {
-		page  int
-		limit int
-	}
-	tests := []struct {
-		name string
-		args args
-		want QueryStakedApplicationsParams
-	}{
-		{"Default Test", args{page: 1, limit: 1}, QueryStakedApplicationsParams{Page: 1, Limit: 1}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := NewQueryStakedApplicationsParams(tt.args.page, tt.args.limit); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewQueryStakedApplicationsParams() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestNewQueryUnstakedApplicationsParams(t *testing.T) {
-	type args struct {
-		page  int
-		limit int
-	}
-	tests := []struct {
-		name string
-		args args
-		want QueryUnstakedApplicationsParams
-	}{
-		{"Default Test", args{page: 1, limit: 1}, QueryUnstakedApplicationsParams{Page: 1, Limit: 1}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := NewQueryUnstakedApplicationsParams(tt.args.page, tt.args.limit); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewQueryUnstakedApplicationsParams() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestNewQueryUnstakingApplicationsParams(t *testing.T) {
-	type args struct {
-		page  int
-		limit int
-	}
-	tests := []struct {
-		name string
-		args args
-		want QueryUnstakingApplicationsParams
-	}{
-		{"Default Test", args{page: 1, limit: 1}, QueryUnstakingApplicationsParams{Page: 1, Limit: 1}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := NewQueryUnstakingApplicationsParams(tt.args.page, tt.args.limit); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewQueryUnstakingApplicationsParams() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestNewQueryApplicationParams(t *testing.T) {
 	type args struct {
 		applicationAddr types.Address