+++ doc/changelog.md
@@ -38,6 +38,7 @@
 - Added configurability for Pocket and Tendermint in a config.json file
 - Patch for Fixed April 17, 2020 consensus failure (0 power consensus failure)
 - Added new endpoints to retrieve transactions from a block (/v1/query/blocktxs) or a new sender (/v1/query/accounttxs)
+- Updated RPC call for Querying Validators and Apps (See Opts in RPC Spec)
 
 ## RC-0.2.1
 - Add version command to CLI